requirements.txt
# requirements.txt
flask==2.3.3
mysql-connector-python==8.2.0
numpy==1.26.3
pandas==2.1.4
beautifulsoup4==4.12.2
selenium==4.16.0
python-dotenv==1.0.0
plotly==5.17.0
scikit-learn==1.3.2
seaborn==0.13.0
matplotlib==3.8.2
logging